requirements.txt
gymnasium
gymnasium[box2d]          # LunarLander-v3 (box2d, pygame, swig)
stable-baselines3
groq                      # LLM backend (llama-3.3-70b-versatile, 1K RPD free)
python-dotenv
numpy
scipy
# google-generativeai     # legacy Gemini backend (deprecated, replaced by groq)
